Removed files and warnings

+++ src/CardManager.cpp
@@ -160,17 +160,17 @@ namespace CardManager {
 	void Initialize() {
 		selectedCard = nullptr;								// Deselect held card, if any
 
-		handBackground.size.x = AEGfxGetWinMaxX();			// The width of the hand BG default to half the screen width
-		handBackground.size.y = AEGfxGetWinMaxY() * 0.25;	// and 25% of the screen height
+		handBackground.size.x = (float)AEGfxGetWinMaxX();			// The width of the hand BG default to half the screen width
+		handBackground.size.y = (float)AEGfxGetWinMaxY() * 0.25f;	// and 25% of the screen height
 
-		handBackground.pos.x = -handBackground.size.x / 2.0;					 // The hand BG should be at the middle
-		handBackground.pos.y = AEGfxGetWinMinY() * 0.95 + handBackground.size.y; // and bottom of the screen
+		handBackground.pos.x = -handBackground.size.x / 2.0f;					 // The hand BG should be at the middle
+		handBackground.pos.y = (float)AEGfxGetWinMinY() * 0.95f + handBackground.size.y; // and bottom of the screen
 
-		cardSpacing = AEGfxGetWinMaxX() * 0.025;			// The gap between cards should at least be 2.5% of the screen width apart
+		cardSpacing = (float)AEGfxGetWinMaxX() * 0.025f;			// The gap between cards should at least be 2.5% of the screen width apart
 
-		cardPositionTemplate.size.y = handBackground.size.y * 0.9;				 // Keep the height of each card to 90% of the hand BG
-		cardPositionTemplate.size.x = cardPositionTemplate.size.y * 0.75;		 // Card width to height ratio is 3:4
-		cardPositionTemplate.pos.y  = handBackground.pos.y - (handBackground.size.y - cardPositionTemplate.size.y) / 2.0; // Set the y-pos so the card is in the middle of hand BG
+		cardPositionTemplate.size.y = handBackground.size.y * 0.9f;				 // Keep the height of each card to 90% of the hand BG
+		cardPositionTemplate.size.x = cardPositionTemplate.size.y * 0.75f;		 // Card width to height ratio is 3:4
+		cardPositionTemplate.pos.y  = handBackground.pos.y - (handBackground.size.y - cardPositionTemplate.size.y) / 2.0f; // Set the y-pos so the card is in the middle of hand BG
 
 		// Initialize card information box
 		cardInfoBox = InfoBox();
